python-graphql-backend/requirements.txt
fastapi
uvicorn
strawberry-graphql
python-dotenv
supabase
